+++ src/GRA.Data/Repository/VendorCodeRepository.cs
@@ -311,12 +311,23 @@ namespace GRA.Data.Repository
                     _.Id,
                     _.IsDeleted,
                     _.FirstName,
-                    _.LastName
+                    _.LastName,
+                    _.ProgramId
                 });
+
+            var groups = await _context
+                .GroupInfos
+                .AsNoTracking()
+                .ToDictionaryAsync(k => k.UserId, v => v.Name);
+
+            var programs = await _context
+                .Programs
+                .AsNoTracking()
+                .ToDictionaryAsync(k => k.Id, v => v.Name);
 
             foreach (var code in codes.Where(_ => _.UserId.HasValue))
             {
-                var user = users.SingleOrDefault(_ => _.Id == code.UserId);
+                var user = await users.SingleOrDefaultAsync(_ => _.Id == code.UserId);
                 if (user == null)
                 {
                     code.IsUserValid = false;
@@ -326,6 +337,12 @@ namespace GRA.Data.Repository
                     code.IsUserValid = !user.IsDeleted;
                     code.FirstName = user.FirstName;
                     code.LastName = user.LastName;
+                    code.ProgramName = programs.TryGetValue(user.ProgramId, out string programValue)
+                        ? programValue
+                        : null;
+                    code.GroupName = groups.TryGetValue(user.Id, out string groupValue)
+                        ? groupValue
+                        : null;
                 }
             }